move `api-keys` endpoints to their own controller
