+++ packages/cli/src/controllers/__tests__/me.controller.test.ts
@@ -2,14 +2,11 @@ import { UserUpdateRequestDto } from '@n8n/api-types';
 import type { Response } from 'express';
 import { mock, anyObject } from 'jest-mock-extended';
 import jwt from 'jsonwebtoken';
-import { randomString } from 'n8n-workflow';
 import { Container } from 'typedi';
 
 import { AUTH_COOKIE_NAME } from '@/constants';
 import { MeController } from '@/controllers/me.controller';
-import type { ApiKey } from '@/databases/entities/api-key';
 import type { User } from '@/databases/entities/user';
-import { ApiKeyRepository } from '@/databases/repositories/api-key.repository';
 import { AuthUserRepository } from '@/databases/repositories/auth-user.repository';
 import { InvalidAuthTokenRepository } from '@/databases/repositories/invalid-auth-token.repository';
 import { UserRepository } from '@/databases/repositories/user.repository';
@@ -21,7 +18,6 @@ import type { PublicUser } from '@/interfaces';
 import { License } from '@/license';
 import { MfaService } from '@/mfa/mfa.service';
 import type { AuthenticatedRequest, MeRequest } from '@/requests';
-import { API_KEY_PREFIX } from '@/services/public-api-key.service';
 import { UserService } from '@/services/user.service';
 import { mockInstance } from '@test/mocking';
 import { badPasswords } from '@test/test-data';
@@ -34,7 +30,6 @@ describe('MeController', () => {
 	const userService = mockInstance(UserService);
 	const userRepository = mockInstance(UserRepository);
 	const mockMfaService = mockInstance(MfaService);
-	const apiKeysRepository = mockInstance(ApiKeyRepository);
 	mockInstance(AuthUserRepository);
 	mockInstance(InvalidAuthTokenRepository);
 	mockInstance(License).isWithinUsersLimit.mockReturnValue(true);
@@ -411,70 +406,6 @@ describe('MeController', () => {
 			};
 
 			await expect(controller.storeSurveyAnswers(req)).rejects.toThrowError(BadRequestError);
-		});
-	});
-
-	describe('API Key methods', () => {
-		let req: AuthenticatedRequest;
-		beforeAll(() => {
-			req = mock<AuthenticatedRequest>({ user: mock<User>({ id: '123' }) });
-		});
-
-		describe('createAPIKey', () => {
-			it('should create and save an API key', async () => {
-				const apiKeyData = {
-					id: '123',
-					userId: '123',
-					label: 'My API Key',
-					apiKey: `${API_KEY_PREFIX}${randomString(42)}`,
-					createdAt: new Date(),
-				} as ApiKey;
-
-				apiKeysRepository.upsert.mockImplementation();
-
-				apiKeysRepository.findOneByOrFail.mockResolvedValue(apiKeyData);
-
-				const newApiKey = await controller.createAPIKey(req);
-
-				expect(apiKeysRepository.upsert).toHaveBeenCalled();
-				expect(apiKeyData).toEqual(newApiKey);
-			});
-		});
-
-		describe('getAPIKeys', () => {
-			it('should return the users api keys redacted', async () => {
-				const apiKeyData = {
-					id: '123',
-					userId: '123',
-					label: 'My API Key',
-					apiKey: `${API_KEY_PREFIX}${randomString(42)}`,
-					createdAt: new Date(),
-				} as ApiKey;
-
-				apiKeysRepository.findBy.mockResolvedValue([apiKeyData]);
-
-				const apiKeys = await controller.getAPIKeys(req);
-				expect(apiKeys[0].apiKey).not.toEqual(apiKeyData.apiKey);
-				expect(apiKeysRepository.findBy).toHaveBeenCalledWith({ userId: req.user.id });
-			});
-		});
-
-		describe('deleteAPIKey', () => {
-			it('should delete the API key', async () => {
-				const user = mock<User>({
-					id: '123',
-					password: 'password',
-					authIdentities: [],
-					role: 'global:member',
-					mfaEnabled: false,
-				});
-				const req = mock<MeRequest.DeleteAPIKey>({ user, params: { id: user.id } });
-				await controller.deleteAPIKey(req);
-				expect(apiKeysRepository.delete).toHaveBeenCalledWith({
-					userId: req.user.id,
-					id: req.params.id,
-				});
-			});
 		});
 	});
 });
