+++ packages/cli/test/integration/me.api.test.ts
@@ -3,55 +3,23 @@ import type { IPersonalizationSurveyAnswersV4 } from 'n8n-workflow';
 import { Container } from 'typedi';
 import validator from 'validator';
 
-import type { ApiKey } from '@/databases/entities/api-key';
 import type { User } from '@/databases/entities/user';
-import { ApiKeyRepository } from '@/databases/repositories/api-key.repository';
 import { ProjectRepository } from '@/databases/repositories/project.repository';
 import { UserRepository } from '@/databases/repositories/user.repository';
-import { PublicApiKeyService } from '@/services/public-api-key.service';
 import { mockInstance } from '@test/mocking';
 
 import { SUCCESS_RESPONSE_BODY } from './shared/constants';
-import { createOwnerWithApiKey, createUser, createUserShell } from './shared/db/users';
+import { createUser, createUserShell } from './shared/db/users';
 import { randomEmail, randomName, randomValidPassword } from './shared/random';
 import * as testDb from './shared/test-db';
 import type { SuperAgentTest } from './shared/types';
 import * as utils from './shared/utils/';
 
 const testServer = utils.setupTestServer({ endpointGroups: ['me'] });
-let publicApiKeyService: PublicApiKeyService;
-
-beforeAll(() => {
-	publicApiKeyService = Container.get(PublicApiKeyService);
-});
 
 beforeEach(async () => {
 	await testDb.truncate(['User']);
 	mockInstance(GlobalConfig, { publicApi: { disabled: false } });
-});
-
-describe('When public API is disabled', () => {
-	let owner: User;
-	let authAgent: SuperAgentTest;
-
-	beforeEach(async () => {
-		owner = await createOwnerWithApiKey();
-
-		authAgent = testServer.authAgentFor(owner);
-		mockInstance(GlobalConfig, { publicApi: { disabled: true } });
-	});
-
-	test('POST /me/api-keys should 404', async () => {
-		await authAgent.post('/me/api-keys').expect(404);
-	});
-
-	test('GET /me/api-keys should 404', async () => {
-		await authAgent.get('/me/api-keys').expect(404);
-	});
-
-	test('DELETE /me/api-key/:id should 404', async () => {
-		await authAgent.delete(`/me/api-keys/${1}`).expect(404);
-	});
 });
 
 describe('Owner shell', () => {
@@ -155,58 +123,6 @@ describe('Owner shell', () => {
 
 			expect(storedShellOwner.personalizationAnswers).toEqual(validPayload);
 		}
-	});
-
-	test('POST /me/api-keys should create an api key', async () => {
-		const newApiKeyResponse = await authOwnerShellAgent.post('/me/api-keys');
-
-		const newApiKey = newApiKeyResponse.body.data as ApiKey;
-
-		expect(newApiKeyResponse.statusCode).toBe(200);
-		expect(newApiKey).toBeDefined();
-
-		const newStoredApiKey = await Container.get(ApiKeyRepository).findOneByOrFail({
-			userId: ownerShell.id,
-		});
-
-		expect(newStoredApiKey).toEqual({
-			id: expect.any(String),
-			label: 'My API Key',
-			userId: ownerShell.id,
-			apiKey: newApiKey.apiKey,
-			createdAt: expect.any(Date),
-			updatedAt: expect.any(Date),
-		});
-	});
-
-	test('GET /me/api-keys should fetch the api key redacted', async () => {
-		const newApiKeyResponse = await authOwnerShellAgent.post('/me/api-keys');
-
-		const retrieveAllApiKeysResponse = await authOwnerShellAgent.get('/me/api-keys');
-
-		expect(retrieveAllApiKeysResponse.statusCode).toBe(200);
-
-		expect(retrieveAllApiKeysResponse.body.data[0]).toEqual({
-			id: newApiKeyResponse.body.data.id,
-			label: 'My API Key',
-			userId: ownerShell.id,
-			apiKey: publicApiKeyService.redactApiKey(newApiKeyResponse.body.data.apiKey),
-			createdAt: expect.any(String),
-			updatedAt: expect.any(String),
-		});
-	});
-
-	test('DELETE /me/api-keys/:id should delete the api key', async () => {
-		const newApiKeyResponse = await authOwnerShellAgent.post('/me/api-keys');
-
-		const deleteApiKeyResponse = await authOwnerShellAgent.delete(
-			`/me/api-keys/${newApiKeyResponse.body.data.id}`,
-		);
-
-		const retrieveAllApiKeysResponse = await authOwnerShellAgent.get('/me/api-keys');
-
-		expect(deleteApiKeyResponse.body.data.success).toBe(true);
-		expect(retrieveAllApiKeysResponse.body.data.length).toBe(0);
 	});
 });
 
@@ -317,61 +233,6 @@ describe('Member', () => {
 
 			expect(storedAnswers).toEqual(validPayload);
 		}
-	});
-
-	test('POST /me/api-keys should create an api key', async () => {
-		const newApiKeyResponse = await testServer.authAgentFor(member).post('/me/api-keys');
-
-		expect(newApiKeyResponse.statusCode).toBe(200);
-		expect(newApiKeyResponse.body.data.apiKey).toBeDefined();
-		expect(newApiKeyResponse.body.data.apiKey).not.toBeNull();
-
-		const newStoredApiKey = await Container.get(ApiKeyRepository).findOneByOrFail({
-			userId: member.id,
-		});
-
-		expect(newStoredApiKey).toEqual({
-			id: expect.any(String),
-			label: 'My API Key',
-			userId: member.id,
-			apiKey: newApiKeyResponse.body.data.apiKey,
-			createdAt: expect.any(Date),
-			updatedAt: expect.any(Date),
-		});
-	});
-
-	test('GET /me/api-keys should fetch the api key redacted', async () => {
-		const newApiKeyResponse = await testServer.authAgentFor(member).post('/me/api-keys');
-
-		const retrieveAllApiKeysResponse = await testServer.authAgentFor(member).get('/me/api-keys');
-
-		expect(retrieveAllApiKeysResponse.statusCode).toBe(200);
-
-		expect(retrieveAllApiKeysResponse.body.data[0]).toEqual({
-			id: newApiKeyResponse.body.data.id,
-			label: 'My API Key',
-			userId: member.id,
-			apiKey: publicApiKeyService.redactApiKey(newApiKeyResponse.body.data.apiKey),
-			createdAt: expect.any(String),
-			updatedAt: expect.any(String),
-		});
-
-		expect(newApiKeyResponse.body.data.apiKey).not.toEqual(
-			retrieveAllApiKeysResponse.body.data[0].apiKey,
-		);
-	});
-
-	test('DELETE /me/api-keys/:id should delete the api key', async () => {
-		const newApiKeyResponse = await testServer.authAgentFor(member).post('/me/api-keys');
-
-		const deleteApiKeyResponse = await testServer
-			.authAgentFor(member)
-			.delete(`/me/api-keys/${newApiKeyResponse.body.data.id}`);
-
-		const retrieveAllApiKeysResponse = await testServer.authAgentFor(member).get('/me/api-keys');
-
-		expect(deleteApiKeyResponse.body.data.success).toBe(true);
-		expect(retrieveAllApiKeysResponse.body.data.length).toBe(0);
 	});
 });
