+++ packages/editor-ui/src/api/api-keys.ts
@@ -2,16 +2,16 @@ import type { ApiKey, IRestApiContext } from '@/Interface';
 import { makeRestApiRequest } from '@/utils/apiUtils';
 
 export async function getApiKeys(context: IRestApiContext): Promise<ApiKey[]> {
-	return await makeRestApiRequest(context, 'GET', '/me/api-keys');
+	return await makeRestApiRequest(context, 'GET', '/api-keys');
 }
 
 export async function createApiKey(context: IRestApiContext): Promise<ApiKey> {
-	return await makeRestApiRequest(context, 'POST', '/me/api-keys');
+	return await makeRestApiRequest(context, 'POST', '/api-keys');
 }
 
 export async function deleteApiKey(
 	context: IRestApiContext,
 	id: string,
 ): Promise<{ success: boolean }> {
-	return await makeRestApiRequest(context, 'DELETE', `/me/api-keys/${id}`);
+	return await makeRestApiRequest(context, 'DELETE', `/api-keys/${id}`);
 }
